+++ bookops_worldcat/query.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
 
-"""
-Handles actual requests to OCLC services
-"""
+"""Handles requests to OCLC web services."""
+
 from __future__ import annotations
-from typing import Optional, Union, Tuple, TYPE_CHECKING
+from typing import Union, Tuple, TYPE_CHECKING
 import sys
 
 from requests.models import PreparedRequest
@@ -17,35 +16,40 @@ if TYPE_CHECKING:
 
 
 class Query:
-    """
-    Sends a request to OCLC service and unifies exceptions.
-    Query object handles refreshing expired token before request is
-    made to the web service.
+    """Sends a request to OCLC web service and unifies exceptions.
 
-    `Query.response` attribute is `requests.Response` instance that
-    can be parsed to extract information received from the web service.
-
+    Query object handles automatic refresh of expired token before each
+    request is made to the web service. `Query.response` attribute is
+    [`requests.Response`](https://requests.readthedocs.io/en/latest/api/#requests.PreparedRequest)
+    instance that can be parsed to extract information received from the web service.
     """
 
     def __init__(
         self,
         session: MetadataSession,
         prepared_request: PreparedRequest,
-        timeout: Optional[
-            Union[int, float, Tuple[int, int], Tuple[float, float]]
-        ] = None,
+        timeout: Union[int, float, Tuple[int, int], Tuple[float, float], None] = (
+            5,
+            5,
+        ),
     ) -> None:
         """Initializes Query object.
 
         Args:
-            session:                        `metadata_api.MetadataSession` instance
-            prepared_request:               `requests.models.PreparedRequest` instance
-            timeout:                        how long to wait for server to send data
-                                            before giving up
-
+            session:
+                `metadata_api.MetadataSession` instance.
+            prepared_request:
+                `requests.PreparedRequest` instance.
+            timeout:
+                How long to wait for server to send data before giving up. Accepts
+                separate values for connect and read timeouts or a single value.
 
         Raises:
-            WorldcatRequestError: If the request encounters an error
+            WorldcatRequestError:
+                If the request encounters any errors.
+            TypeError:
+                If `prepared_request` arg is passed anything other than a
+                `requests.PreparedRequest` object.
         """
         if not isinstance(prepared_request, PreparedRequest):
             raise TypeError("Invalid type for argument 'prepared_request'.")